LIBITD-2566. Updated styling.
The styling was updated to mirror Grove and track the original
IPManager more closely. I also modified how the homepage works -
the homepage now links to the group list view while the django
administration is accessable via a nav link.

This UI change is dependent on styles added to
django-umd-lib-style.

https://umd-dit.atlassian.net/browse/LIBITD-2566